Server/requirements.txt
fastapi
uvicorn
numpy
opencv-python
pillow
python-multipart
scipy